Move login routes to its own controller
