+++ src/Application.php
@@ -309,7 +309,9 @@ class Application extends Silex\Application
         }
 
         // Mount the 'backend' on the branding:path setting. Defaults to '/bolt'.
-        $this->mount($this['config']->get('general/branding/path'), new Controllers\Backend());
+        $backendPrefix = $this['config']->get('general/branding/path');
+        $this->mount($backendPrefix, new Controllers\Login());
+        $this->mount($backendPrefix, new Controllers\Backend());
 
         // Mount the 'async' controllers on /async. Not configurable.
         $this->mount('/async', new Controllers\Async());
